# services/tanren-api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "tanren",
    "description": "Tanren worker-manager HTTP API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/run/provision": {
      "post": {
        "tags": [
          "run"
        ],
        "summary": "Run Provision",
        "description": "Provision a remote execution environment.",
        "operationId": "run_provision_api_v1_run_provision_post",
        "parameters": [
          {
            "name": "x-api-key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "title": "X-Api-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProvisionRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunEnvironment"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/run/{env_id}/execute": {
      "post": {
        "tags": [
          "run"
        ],
        "summary": "Run Execute",
        "description": "Execute a phase against a provisioned environment.",
        "operationId": "run_execute_api_v1_run__env_id__execute_post",
        "parameters": [
          {
            "name": "env_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Environment identifier",
              "title": "Env Id"
            },
            "description": "Environment identifier"
          },
          {
            "name": "x-api-key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "title": "X-Api-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExecuteRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunExecuteAccepted"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/run/{env_id}/teardown": {
      "post": {
        "tags": [
          "run"
        ],
        "summary": "Run Teardown",
        "description": "Teardown a provisioned environment.",
        "operationId": "run_teardown_api_v1_run__env_id__teardown_post",
        "parameters": [
          {
            "name": "env_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Environment identifier",
              "title": "Env Id"
            },
            "description": "Environment identifier"
          },
          {
            "name": "x-api-key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunTeardownAccepted"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/run/full": {
      "post": {
        "tags": [
          "run"
        ],
        "summary": "Run Full",
        "description": "Full lifecycle: provision, execute, teardown. Returns ID for polling.",
        "operationId": "run_full_api_v1_run_full_post",
        "parameters": [
          {
            "name": "x-api-key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "title": "X-Api-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RunFullRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DispatchAccepted"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/run/{env_id}/status": {
      "get": {
        "tags": [
          "run"
        ],
        "summary": "Run Status",
        "description": "Poll status of a running environment.",
        "operationId": "run_status_api_v1_run__env_id__status_get",
        "parameters": [
          {
            "name": "env_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Environment identifier",
              "title": "Env Id"
            },
            "description": "Environment identifier"
          },
          {
            "name": "x-api-key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunStatus"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Cli": {
        "type": "string",
        "enum": [
          "opencode",
          "codex",
          "claude",
          "bash"
        ],
        "title": "Cli",
        "description": "CLI tool types from PROTOCOL.md Section 2."
      },
...
      "DispatchAccepted": {
        "properties": {
          "dispatch_id": {
            "type": "string",
            "title": "Dispatch Id",
            "description": "Auto-generated workflow identifier"
          },
          "status": {
            "type": "string",
            "title": "Status",
            "description": "Acceptance status",
            "default": "accepted"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "dispatch_id"
        ],
        "title": "DispatchAccepted",
        "description": "Thin response confirming dispatch acceptance."
      },
...
      "ExecuteRequest": {
        "properties": {
          "project": {
            "type": "string",
            "title": "Project",
            "description": "Project name"
          },
          "spec_path": {
            "type": "string",
            "title": "Spec Path",
            "description": "Spec folder path"
          },
          "phase": {
            "$ref": "#/components/schemas/Phase",
            "description": "Phase to execute"
          },
          "cli": {
            "$ref": "#/components/schemas/Cli",
            "description": "CLI tool",
            "default": "claude"
          },
          "model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model",
            "description": "Model identifier"
          },
          "timeout": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Timeout",
            "description": "Max execution seconds",
            "default": 1800
          },
          "context": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Context",
            "description": "Extra context"
          },
          "gate_cmd": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Gate Cmd",
            "description": "Gate command"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "project",
          "spec_path",
          "phase"
        ],
        "title": "ExecuteRequest",
        "description": "Request body for POST /run/{env_id}/execute."
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "Outcome": {
        "type": "string",
        "enum": [
          "success",
          "fail",
          "blocked",
          "error",
          "timeout"
        ],
        "title": "Outcome",
        "description": "Result outcomes from PROTOCOL.md Section 3."
      },
...
      "Phase": {
        "type": "string",
        "enum": [
          "do-task",
          "audit-task",
          "run-demo",
          "audit-spec",
          "investigate",
          "gate",
          "setup",
          "cleanup"
        ],
        "title": "Phase",
        "description": "Dispatch phase types from PROTOCOL.md Section 2."
      },
...
      "ProvisionRequest": {
        "properties": {
          "project": {
            "type": "string",
            "title": "Project",
            "description": "Project name"
          },
          "branch": {
            "type": "string",
            "title": "Branch",
            "description": "Git branch"
          },
          "environment_profile": {
            "type": "string",
            "title": "Environment Profile",
            "description": "Environment profile",
            "default": "default"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "project",
          "branch"
        ],
        "title": "ProvisionRequest",
        "description": "Request to provision a VM \u2014 subset of Dispatch fields."
      },
...
      "RunEnvironment": {
        "properties": {
          "env_id": {
            "type": "string",
            "title": "Env Id",
            "description": "Environment identifier"
          },
          "vm_id": {
            "type": "string",
            "title": "Vm Id",
            "description": "Backing VM identifier"
          },
          "host": {
            "type": "string",
            "title": "Host",
            "description": "VM hostname or IP"
          },
          "status": {
            "$ref": "#/components/schemas/RunEnvironmentStatus",
            "description": "Environment status",
            "default": "provisioned"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "env_id",
          "vm_id",
          "host"
        ],
        "title": "RunEnvironment",
        "description": "Provisioned run environment handle."
      },
      "RunEnvironmentStatus": {
        "type": "string",
        "enum": [
          "provisioning",
          "provisioned",
          "executing",
          "tearing_down",
          "completed",
          "failed"
        ],
        "title": "RunEnvironmentStatus",
        "description": "Lifecycle status of a run environment."
      },
      "RunExecuteAccepted": {
        "properties": {
          "env_id": {
            "type": "string",
            "title": "Env Id",
            "description": "Environment identifier"
          },
          "dispatch_id": {
            "type": "string",
            "title": "Dispatch Id",
            "description": "Dispatch workflow identifier"
          },
          "status": {
            "$ref": "#/components/schemas/RunEnvironmentStatus",
            "description": "Execution status",
            "default": "executing"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "env_id",
          "dispatch_id"
        ],
        "title": "RunExecuteAccepted",
        "description": "Confirmation that execution was accepted."
      },
      "RunFullRequest": {
        "properties": {
          "project": {
            "type": "string",
            "title": "Project",
            "description": "Project name"
          },
          "branch": {
            "type": "string",
            "title": "Branch",
            "description": "Git branch"
          },
          "spec_path": {
            "type": "string",
            "title": "Spec Path",
            "description": "Spec folder path"
          },
          "phase": {
            "$ref": "#/components/schemas/Phase",
            "description": "Phase to execute"
          },
          "environment_profile": {
            "type": "string",
            "title": "Environment Profile",
            "description": "Environment profile",
            "default": "default"
          },
          "timeout": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Timeout",
            "description": "Max execution seconds",
            "default": 1800
          },
          "context": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Context",
            "description": "Extra context"
          },
          "gate_cmd": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Gate Cmd",
            "description": "Gate command"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "project",
          "branch",
          "spec_path",
          "phase"
        ],
        "title": "RunFullRequest",
        "description": "Full lifecycle request \u2014 combines provision + execute fields."
      },
      "RunStatus": {
        "properties": {
          "env_id": {
            "type": "string",
            "title": "Env Id",
            "description": "Environment identifier"
          },
          "status": {
            "$ref": "#/components/schemas/RunEnvironmentStatus",
            "description": "Current environment status"
          },
          "phase": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Phase"
              },
              {
                "type": "null"
              }
            ],
            "description": "Current phase if executing"
          },
          "outcome": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Outcome"
              },
              {
                "type": "null"
              }
            ],
            "description": "Final outcome if completed"
          },
          "started_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At",
            "description": "ISO 8601 start timestamp"
          },
          "duration_secs": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Duration Secs",
            "description": "Elapsed seconds"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "env_id",
          "status"
        ],
        "title": "RunStatus",
        "description": "Status of a running environment."
      },
      "RunTeardownAccepted": {
        "properties": {
          "env_id": {
            "type": "string",
            "title": "Env Id",
            "description": "Environment identifier"
          },
          "status": {
            "$ref": "#/components/schemas/RunEnvironmentStatus",
            "description": "Teardown status",
            "default": "tearing_down"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "env_id"
        ],
        "title": "RunTeardownAccepted",
        "description": "Confirmation that teardown was accepted."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}